feat: 타임라인 이벤트 timestamp를 hits.time_sec 기반 소수점 초로 변경
- TimelineEventDto.timestamp Integer → Float
- getVideoDetail에서 hits.time_sec 맵 구성 후 hitNumber로 매핑

+++ src/main/java/com/rallytrack/backend/domain/video/dto/VideoDetailResponse.java
@@ -41,7 +41,7 @@ public class VideoDetailResponse {
     @AllArgsConstructor
     public static class TimelineEventDto {
         private Long eventId;
-        private Integer timestamp;
+        private Float timestamp;
         private String displayTime;
         private String type;       // EventType enum의 name() 값
         private String title;
